upgrade remix.init script to esm
